prymal/dashboards/requirements.txt
boto3
botocore
requests
pandas
datetime
plotly.express
plotly.graph_objects
dash
dash_core_components
dash_html_components
